refactor: CommentController.java에서 사용하는 DTO를 PostController에서 분리
